move app update to shell
