+++ pubspec.yaml
@@ -54,7 +54,6 @@ dependencies:
   flutter_mentions: ^2.0.1
   photo_view: ^0.15.0
   qr_flutter: ^4.0.0
-  matomo_tracker: ^5.1.0
   shared_preferences: ^2.3.4
 
   custom_refresh_indicator: ^4.0.1
